+++ packages/camera/camera_android_camerax/lib/src/android_camera_camerax.dart
@@ -914,6 +914,74 @@ class AndroidCameraCameraX extends CameraPlatform {
     }
   }
 
+  @override
+  Future<Iterable<VideoStabilizationMode>> getSupportedVideoStabilizationModes(
+    int cameraId,
+  ) async {
+    return (await _getSupportedVideoStabilizationModeMap(cameraId)).keys;
+  }
+
+  /// Throws a [ArgumentError] when an unsupported [mode] is
+  /// supplied.
+  @override
+  Future<void> setVideoStabilizationMode(
+    int cameraId,
+    VideoStabilizationMode mode,
+  ) async {
+    final Map<VideoStabilizationMode, int> availableModes =
+        await _getSupportedVideoStabilizationModeMap(cameraId);
+
+    final int? controlMode = availableModes[mode];
+    if (controlMode == null) {
+      throw ArgumentError('Unavailable video stabilization mode.', 'mode');
+    }
+
+    final CaptureRequestOptions captureRequestOptions = proxy
+        .newCaptureRequestOptions(
+          options: <CaptureRequestKey, Object?>{
+            proxy.controlVideoStabilizationModeRequest(): controlMode,
+          },
+        );
+
+    final Camera2CameraControl camera2Control = proxy.fromCamera2CameraControl(
+      cameraControl: cameraControl,
+    );
+    await camera2Control.addCaptureRequestOptions(captureRequestOptions);
+  }
+
+  /// Gets a map of video stabilization control modes that are supported for the
+  /// selected camera, indexed by the respective [VideoStabilizationMode].
+  Future<Map<VideoStabilizationMode, int>>
+  _getSupportedVideoStabilizationModeMap(int cameraId) async {
+    final CameraInfo? camInfo = cameraInfo;
+    if (camInfo == null) {
+      return <VideoStabilizationMode, int>{};
+    }
+    final Camera2CameraInfo camera2CameraInfo = proxy.fromCamera2CameraInfo(
+      cameraInfo: cameraInfo!,
+    );
+
+    final List<int> controlModes =
+        (await camera2CameraInfo.getCameraCharacteristic(
+              proxy
+                  .controlAvailableVideoStabilizationModesCameraCharacteristics(),
+            ))!
+            as List<int>;
+
+    final Map<VideoStabilizationMode, int>
+    modes = <VideoStabilizationMode, int>{
+      for (final int controlMode in controlModes)
+        // https://developer.android.com/reference/android/hardware/camera2/CameraMetadata#CONTROL_VIDEO_STABILIZATION_MODE_OFF
+        if (controlMode == 0)
+          VideoStabilizationMode.off: 0
+        // https://developer.android.com/reference/android/hardware/camera2/CameraMetadata#CONTROL_VIDEO_STABILIZATION_MODE_ON
+        else if (controlMode == 1)
+          VideoStabilizationMode.level1: 1,
+    };
+
+    return modes;
+  }
+
   /// The ui orientation changed.
   @override
   Stream<DeviceOrientationChangedEvent> onDeviceOrientationChanged() {