+++ packages/camera/camera_android_camerax/lib/src/camerax_proxy.dart
@@ -66,11 +66,15 @@ class CameraXProxy {
     this.ratio_4_3FallbackAutoStrategyAspectRatioStrategy =
         _ratio_4_3FallbackAutoStrategyAspectRatioStrategy,
     this.controlAELockCaptureRequest = _controlAELockCaptureRequest,
+    this.controlVideoStabilizationModeRequest =
+        _controlVideoStabilizationModeRequest,
     this.infoSupportedHardwareLevelCameraCharacteristics =
         _infoSupportedHardwareLevelCameraCharacteristics,
     this.sensorOrientationCameraCharacteristics =
         _sensorOrientationCameraCharacteristics,
     this.getNv21BufferImageProxyUtils = ImageProxyUtils.getNv21Buffer,
+    this.controlAvailableVideoStabilizationModesCameraCharacteristics =
+        _controlAvailableVideoStabilizationModesCameraCharacteristics,
   });
 
   /// Handles adding support for generic classes.
@@ -369,6 +373,9 @@ class CameraXProxy {
   /// Calls to [CaptureRequest.controlAELock].
   CaptureRequestKey Function() controlAELockCaptureRequest;
 
+  /// Calls to [CaptureRequest.controlVideoStabilizationMode].
+  CaptureRequestKey Function() controlVideoStabilizationModeRequest;
+
   /// Calls to [CameraCharacteristics.infoSupportedHardwareLevel].
   final CameraCharacteristicsKey Function()
   infoSupportedHardwareLevelCameraCharacteristics;
@@ -376,6 +383,10 @@ class CameraXProxy {
   /// Calls to [CameraCharacteristics.sensorOrientation].
   final CameraCharacteristicsKey Function()
   sensorOrientationCameraCharacteristics;
+
+  /// Calls to [CameraCharacteristics.controlAvailableVideoStabilizationModes].
+  final CameraCharacteristicsKey Function()
+  controlAvailableVideoStabilizationModesCameraCharacteristics;
 
   /// Calls to [ImageProxyUtils.getNv21Buffer].
   final Future<Uint8List?> Function(
@@ -407,10 +418,17 @@ class CameraXProxy {
   static CaptureRequestKey _controlAELockCaptureRequest() =>
       CaptureRequest.controlAELock;
 
+  static CaptureRequestKey _controlVideoStabilizationModeRequest() =>
+      CaptureRequest.controlVideoStabilizationMode;
+
   static CameraCharacteristicsKey
   _infoSupportedHardwareLevelCameraCharacteristics() =>
       CameraCharacteristics.infoSupportedHardwareLevel;
 
   static CameraCharacteristicsKey _sensorOrientationCameraCharacteristics() =>
       CameraCharacteristics.sensorOrientation;
+
+  static CameraCharacteristicsKey
+  _controlAvailableVideoStabilizationModesCameraCharacteristics() =>
+      CameraCharacteristics.controlAvailableVideoStabilizationModes;
 }