+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/Mocks/MockCaptureDevice.swift
@@ -112,6 +112,11 @@ class MockCaptureDevice: NSObject, FLTCaptureDevice {
   func iso() -> Float {
     return 0
   }
+  
+  func isVideoStabilizationModeSupported(_ videoStabilizationMode: AVCaptureVideoStabilizationMode) -> Bool {
+    return false
+  }
+
 
   func lockForConfiguration() throws {
     try lockForConfigurationStub?()