+++ packages/camera/camera/CHANGELOG.md
@@ -3,6 +3,10 @@
 * Updates minimum supported SDK version to Flutter 3.29/Dart 3.7.
 * Updates README to reflect that only Android API 24+ is supported.
 
+## 0.12.0
+
+* Adds support for video stabilization.
+
 ## 0.11.2
 
 * Fixes overflowed toggles in the camera example.